refactor: AppConfig class 에서 @ComponentScan 으로 Bean제작 방식 변경
