+++ src/main/java/com/example/voucher_manager/VoucherManagerApplication.java
@@ -1,5 +1,6 @@
 package com.example.voucher_manager;
 
+import com.example.voucher_manager.application.RunApplication;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
