+++ src/main/java/com/example/voucher_manager/application/ApplicationStatus.java
@@ -1,4 +1,4 @@
-package com.example.voucher_manager;
+package com.example.voucher_manager.application;
 
 public enum ApplicationStatus {
     IDLE,
