+++ src/main/java/com/example/voucher_manager/domain/repository/MemoryVoucherRepository.java
@@ -1,10 +1,14 @@
 package com.example.voucher_manager.domain.repository;
 
 import com.example.voucher_manager.domain.voucher.Voucher;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Repository;
 
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 
+@Repository
+@Profile("dev")
 public class MemoryVoucherRepository implements VoucherRepository {
     private final Map<UUID, Voucher> memory = new ConcurrentHashMap<>();
