+++ src/main/java/com/example/voucher_manager/domain/service/CustomerService.java
@@ -2,9 +2,11 @@ package com.example.voucher_manager.domain.service;
 
 import com.example.voucher_manager.domain.customer.Customer;
 import com.example.voucher_manager.domain.repository.CustomerRepository;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Service
 public class CustomerService {
     private final CustomerRepository customerRepository;
