+++ src/main/java/com/example/voucher_manager/domain/voucher/Voucher.java
@@ -1,8 +1,9 @@
 package com.example.voucher_manager.domain.voucher;
 
+import java.io.Serializable;
 import java.util.UUID;
 
-public interface Voucher {
+public interface Voucher extends Serializable {
     UUID getVoucherId();
     Long discount(Long regularPrice);
 }
