+++ src/main/java/com/example/voucher_manager/io/ConsoleInput.java
@@ -1,7 +1,10 @@
 package com.example.voucher_manager.io;
 
+import org.springframework.stereotype.Component;
+
 import java.util.Scanner;
 
+@Component
 public class ConsoleInput implements Input{
     private static final Scanner scanner = new Scanner(System.in);
